pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentverse"
version = "0.1.0"
description = "A decentralized marketplace for AI agents"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "streamlit>=1.31.0",
    "web3>=6.11.1",
    "ipfshttpclient>=0.8.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.4",
    "numpy>=1.24.3",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "aiohttp>=3.9.1",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=0.900",
    "flake8>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=agentverse"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 